refractor: pagination component
