--- next-react-query-tailwind/src/components/IssueFilters/Pagination.module.css
@@ -1,21 +1,0 @@
-@tailwind utilities;
-
-.container {
-  @apply flex items-center justify-center mt-4;
-}
-
-.group {
-  @apply relative z-0 inline-flex shadow-sm rounded-md;
-}
-
-.button {
-  @apply relative inline-flex items-center px-4 py-2 border border-gray-300 bg-gray-100 text-sm font-semibold text-blue-500 hover:bg-blue-500 hover:border-blue-500  hover:text-white focus:z-10 focus:outline-none focus:ring-1 focus:ring-indigo-500 focus:border-indigo-500 disabled:opacity-50 disabled:cursor-default disabled:hover:bg-gray-100 disabled:hover:text-blue-500 disabled:hover:border-gray-200 transition-colors duration-150 ease-in-out;
-}
-
-.buttonPrev {
-  @apply rounded-l-lg;
-}
-
-.buttonNext {
-  @apply rounded-r-lg;
-}
