+++ next-react-query-tailwind/src/components/IssueFilters/index.ts
@@ -1,3 +1,2 @@
 export { default as IssueFilters } from './IssueFilters';
-export { default as Pagination } from './Pagination';
 export * from './useIssueFilters';
